Move Natural tests into the Module tests.
